functions/AddRagData/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai
langchain
langchain_openai
langchain_community
pymongo
bson

aiofiles==24.1.0
    # via quart
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==24.3.0
    # via aiohttp
blinker==1.9.0
    # via
    #   flask
    #   quart
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   flask
    #   quart
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via pymongo
exceptiongroup==1.2.2
    # via
    #   anyio
    #   hypercorn
    #   taskgroup
flask==3.1.0
    # via quart
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-community
hypercorn==0.17.3
    # via
    #   quart
    #   quartapp (src/pyproject.toml)
hyperframe==6.0.1
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via
    #   flask
    #   quart
jinja2==3.1.5
    # via
    #   flask
    #   quart
jiter==0.8.2
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.14
    # via
    #   langchain-community
    #   quartapp (src/pyproject.toml)
langchain-community==0.3.14
    # via quartapp (src/pyproject.toml)
langchain-core==0.3.29
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.2.14
    # via quartapp (src/pyproject.toml)
langchain-text-splitters==0.3.4
    # via langchain
langsmith==0.2.10
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==3.0.2
    # via
    #   jinja2
    #   quart
    #   werkzeug
marshmallow==3.23.3
    # via dataclasses-json
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
openai==1.59.3
    # via langchain-openai
orjson==3.10.13
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   marshmallow
priority==2.0.0
    # via hypercorn
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.10.4
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via langchain-community
pymongo==4.10.1
    # via quartapp (src/pyproject.toml)
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   quartapp (src/pyproject.toml)
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
quart==0.20.0
    # via quartapp (src/pyproject.toml)
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.36
    # via
    #   langchain
    #   langchain-community
taskgroup==0.2.2
    # via hypercorn
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.8.0
    # via langchain-openai
tomli==2.2.1
    # via hypercorn
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   hypercorn
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   taskgroup
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.3.0
    # via requests
werkzeug==3.1.3
    # via
    #   flask
    #   quart
wsproto==1.2.0
    # via hypercorn
yarl==1.18.3
    # via aiohttp